% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/app_ESR.R
\name{app_ESR}
\alias{app_ESR}
\title{Run ESR shiny apps}
\usage{
app_ESR(app, ...)
}
\arguments{
\item{app}{\code{\link{character}} (required): name of the ESR app to start}

\item{...}{further arguments to pass to \code{\link{runApp}}}
}
\value{
#
}
\description{
A wrapper for shiny::runApp() to start interactive shiny apps
}
\details{
This is a wrapper function for \code{runApp}.
}
\note{
#
}
\examples{
\dontrun{
app_ESR("converter")
app_ESR("spectrum")
}
}
\author{
Christoph Burow, University of Cologne (Germany)
}
\references{
In progress
}
\seealso{
\code{\link{runApp}}
}

