% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{as.ESR.Spectrum}
\alias{as.ESR.Spectrum}
\title{Coercion from and to ESR.Spectrum}
\usage{
as.ESR.Spectrum(x)
}
\arguments{
\item{x}{an object}
}
\value{
\code{as.ESR.Spectrum} returns an \code{ESR.Spectrum} object.
}
\description{
A generic function \code{as.ESR.Spectrum} for coercing objects
to class "\code{ESR.Spectrum}".
}
\details{
\code{as.ESR.Spectrum} currently includes methods for \link{data.frame} objects.
}
\examples{
## TODO
}
\author{
Christoph Burow, University of Cologne (Germany)
}
\seealso{
\link{R6}, \link{R6Class}
}

