% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESR-package.R
\docType{package}
\name{ESR-package}
\alias{ESR-package}
\title{Collection of functions for Electron Spin Resonance Dating data analysis}
\description{
This package provides various functions developed for the purpose of
Electron Spin Resonance Dating data analysis.
}
\details{
\tabular{ll}{ Package: \tab ESR\cr 
Type: \tab Package\cr 
Version: 0.1.0.9031\tab \cr 
Date: \tab 2015-07-03\cr 
Author: \tab Christoph Burow \cr
Maintainer: \tab Christoph Burow <christoph.burow@gmx.net>\cr 
License: \tab GPL-3\cr}
}
\author{
Christoph Burow (University of Cologne, Germany)
}
\keyword{package}
