% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESR-class.R
\docType{class}
\name{ESR}
\alias{ESR}
\title{ESR class objects}
\format{\link{R6Class} generator objects}
\usage{
# ESR$new(originator, data)
# ESR$set_data(data)
# ESR$set_origin(originator)
}
\arguments{
\item{originator}{\code{\link{character}} name of the function call that
created this object}

\item{data}{\code{\link{data.frame}} the ESR spectrum to be stored in the}
}
\description{
Objects of class \code{ESR}, \code{ESR.Spectrum} are \link{R6} objects and hence are 
usable in a more typical object-oriented language than R. For the user the main difference 
to R's internal \code{S3} and \code{S4} object implementation is that methods for \link{R6} 
objects belong to the object itself. See details.
}
\section{Slots}{

\describe{
\item{\code{originator}}{\code{\link{character}} Name of the function call that
created this object}

\item{\code{data}}{\code{\link{data.frame}} XY data of the ESR spectrum}
}}
\section{Methods}{

\describe{
  \item{\code{new(originator, data)}}{ Initialize a new object of class \code{ESR}}
  \item{\code{set_data(data)}}{ Sets the data of the object }
  \item{\code{set_origin(originator)}}{ Sets the originator of the object }
}
}
\examples{

## create an empty ESR base class object
x <- ESR$new()

}
\author{
Christoph Burow, University of Cologne (Germany)
}
\keyword{classes}
\keyword{methods}

