% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESR-class.R
\docType{class}
\name{ESR.Spectrum.2D}
\alias{ESR.Spectrum.2D}
\title{ESR.Spectrum.2D class objects}
\format{\link{R6Class} generator objects}
\usage{
# ESR.Spectrum.2D$new()
}
\description{
Objects of class \code{ESR}, \code{ESR.Spectrum} are \link{R6} objects and hence are 
usable in a more typical object-oriented language than R. For the user the main difference 
to R's internal \code{S3} and \code{S4} object implementation is that methods for \link{R6} 
objects belong to the object itself. See details.
}
\section{Methods}{


See \code{ESR.Spectrum}-class
}

\examples{

# none

}
\author{
Christoph Burow, University of Cologne (Germany)
}
\keyword{classes}
\keyword{methods}
