% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESR-class.R
\docType{class}
\name{ESR.Spectrum}
\alias{ESR.Spectrum}
\title{ESR.Spectrum class objects}
\format{\link{R6Class} generator objects}
\usage{
# ESR.Spectrum$new(originator, data)
# ESR.Spectrum$set_origin(originator)
# ESR.Spectrum$set_data(data)
# ESR.Spectrum$set_par(par)
# ESR.Spectrum$get_diff(x, order, ...)
# ESR.Spectrum$get_integral(x)
# ESR.Spectrum$get_spline(x, ...)
# ESR.Spectrum$get_gvalues(v, H, x)
# ESR.Spectrum$get_peaks(x, interval, th = 10)
}
\arguments{
\item{x}{\code{\link{data.frame}} XY data of the ESR spectrum}

\item{originator}{\code{\link{character}} name of the function call that
created this object}

\item{data}{\code{\link{data.frame}} XY values of the ESR spectrum}

\item{par}{\code{\link{data.frame}} experimental parameters}

\item{order}{\code{\link{integer}} order of the differential}

\item{v}{\code{\link{numeric}} The microwave frequency in GHz}

\item{H}{\code{\link{numeric}} A vector of magnetic field values in Gauss}

\item{interval}{\code{\link{numeric}} a vector of length two specifying
the range of x-values where peaks are searched}

\item{th}{\code{\link{numeric}}: an integer specifying the number of neighbouring values
to compare each x-value to}

\item{...}{Further arguments passed to method specific functions.}
}
\description{
Objects of class \code{ESR}, \code{ESR.Spectrum} are \link{R6} objects and hence are 
usable in a more typical object-oriented language than R. For the user the main difference 
to R's internal \code{S3} and \code{S4} object implementation is that methods for \link{R6} 
objects belong to the object itself. See details.
}
\section{Slots}{

\describe{
\item{\code{originator}}{\code{\link{character}} Name of the function call that
created this object}

\item{\code{data}}{\code{\link{data.frame}} XY data of the ESR spectrum}

\item{\code{parameter}}{\code{\link{data.frame}} experimental parameters}

\item{\code{type}}{\code{\link{character}} currently unused}
}}

\section{Methods}{

\describe{
  \item{\code{new(originator, data)}}{ Initialize a new object of class \code{ESR.Spectrum}}
  \item{\code{set_data(data)}}{ Sets the data of the object }
  \item{\code{set_origin(originator)}}{ Sets the originator of the object }
  \item{\code{set_par(par)}}{ Sets the experimental settings of the object }
  \item{\code{get_diff(x, order = 1, ...)}}{ Returns the differential of the spectrum }
  \item{\code{get_integral(x)}}{ Returns the integrant of the spectrum }
  \item{\code{get_spline(x, ...)}}{ Returns a smoothed spectrum using the \code{\link{smooth.spline}} function }
  \item{\code{get_gvalues(v, H, x)}}{ Calculates and returns the g-values of the spectrum }
  \item{\code{get_peaks(x, interval, th = 10)}}{ Returns a \code{\link{vector}} of local minima and maxima in the spectrum. For details see \code{\link{find_Peaks}}. }
}
}

\examples{

## create an object of class ESR.Spectrum
## and fill random data
x <- ESR.Spectrum$new()
x$set_data(data.frame(seq(3350, 3450, length.out = 1024), 
                                runif(1024, -1000, 1000)))

# plot the random data
plot(x)

\dontrun{
## public set methods
x$set_par()
x$set_data()
x$set_origin()
}

## public get methods
x$get_diff()
x$get_integral()
x$get_spline()
\dontrun{
x$get_gvalues()
}
}
\author{
Christoph Burow, University of Cologne (Germany)
}
\keyword{classes}
\keyword{methods}
