\name{ExampleData.dpph}
\alias{ExampleData.dpph}
\docType{data}
\title{
ESR spectrum of the dpph standard
}
\description{
ESR spectrum of the dpph standard measured at 100 K
}
\usage{data(ExampleData.dpph)}
\format{
  A data frame with 1024 observations on the following 2 variables.
  \describe{
    \item{\code{magnetic.field}}{magnetic field (G)}
    \item{\code{ESR.intensity}}{ESR intensity (a.u.)}
  }
}
\details{
In progress
}
\source{
In progress
}
\references{
In progress
}
\examples{
data(ExampleData.dpph)
plot(ExampleData.dpph)
}
\keyword{datasets}
