\name{ExampleData.ESRspectra}
\alias{ExampleData.ESRspectra}
\docType{data}
\title{
ESR spectra of a mollusc and the dpph standard
}
\description{
ESR spectra of a mollusc from the foothills of the Alps and the dpph standard measured at 100 K
}
\usage{ExampleData.ESRspectra}
\format{
  A list of ESR spectra where each element contains 1024 observations on the following 2 variables.
}
\details{
In progress
}
\source{
In progress
}
\references{
In progress
}
\examples{
data(ExampleData.ESRspectra)
}
\keyword{datasets}
