% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/le_toucan.R
\name{le_toucan}
\alias{le_toucan}
\title{Le Toucan has arrived}
\usage{
le_toucan(...)
}
\arguments{
\item{...}{le toucan takes not one, but all of the arguments. The more, the
better.}
}
\value{
Unleashes le toucan to the world.
}
\description{
PRAISE!
}
\examples{

le_toucan()

}
