% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DRC.R
\name{plot_DRC}
\alias{plot_DRC}
\title{Plot a dose response curve}
\usage{
plot_DRC(object, interactive = FALSE, ...)
}
\arguments{
\item{object}{\code{\link{list}} (required): a results object returned from
the \code{fit_DRC} function}

\item{interactive}{\code{link{logical}} (optional): create an interactive plot
(requires the plotly package)}

\item{...}{further graphical parameters to be passed to \code{plot}}
}
\value{
Returns a plot
}
\description{
This functions plots the dose response curve of an object that
was created with the fit_DRC() function
}
\examples{

#' ##load example data
data(ExampleData.De, envir = environment())
obj <- fit_DRC(input.data = ExampleData.De, fit.weights = 'prop')
plot_DRC(obj)

}
