% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_ESRspectrum.R
\name{read_ESRspectrum}
\alias{read_ESRspectrum}
\title{Read in multiple associated ESR spectra}
\usage{
read_ESRspectrum(file, ...)
}
\arguments{
\item{file}{\code{\link{character}} (required): directory where the
spectra files are stored. Alternatively, a compressed zip file can be
provided. See details!}

\item{...}{further arguments}
}
\value{
Returns a terminal output. In addition an
\code{\linkS4class{RLum.Results}} object is returned. \cr\cr The output
should be accessed using the function \code{\link{get_RLum.Results}}
}
\description{
Use this function to import a series of associated ESR spectra into R.
}
\details{
This is a wrapper function for \code{read.table}. The function should be
used to read in a series of associated ESR spectrum files. A list with all
spectrum data is returned, which can be passed to \code{plot_ESRspectrum}
for plotting the spectra. \cr\cr \bold{Compressed zip archives} \cr For
\code{file} a zip archive can be provided. The file has to have a .zip
file extension and must contain all spectrum files in no subfile.
}
\note{
#
}
\examples{
# Import ASCII text file
file1 <- system.file("extdata", "coral.txt", package = "ESR")
spec1 <- read_ESRspectrum(file1)

# Import .zip archive
file2 <- system.file("extdata", "mollusc.zip", package = "ESR")
spec2 <- read_ESRspectrum(file2)

# Import Bruker ESP300-E raw binary spectrum
file3 <- system.file("extdata", "mollusc.SPC", package = "ESR")
spec3 <- read_ESRspectrum(file3)

# Import Bruker ELEXSYS500 spectrum (ASCII)
file4 <- system.file("extdata", "dpph.ASC", package = "ESR")
spec4 <- read_ESRspectrum(file4)

# Import Bruker ELEXSYS500 raw binary spectrum
file5 <- system.file("extdata", "quartz.DTA", package = "ESR")
spec5 <- read_ESRspectrum(file5)

# Import all example data sets at once by providing only the directory
dir <- system.file("extdata", package = "ESR")
specs <- read_ESRspectrum(dir)
}
\author{
Christoph Burow, University of Cologne (Germany)
}
\references{
In progress
}
\seealso{
\code{\link{read.table}}, \code{\link{unzip}}, \code{\link{unz}}
}

