% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R, R/ridge_fit.R
\name{ridge_fit}
\alias{ridge_fit}
\title{Fit a Ridge regression model}
\usage{
ridge_fit(X, y, lambda = 1)

ridge_fit(X, y, lambda = 1)
}
\arguments{
\item{X}{A numeric matrix of predictors.}

\item{y}{A numeric response vector.}

\item{lambda}{A non-negative regularization parameter.}
}
\value{
A vector containing the estimated Ridge coefficients.

A vector containing the estimated Ridge coefficients.
}
\description{
Fits a Ridge regression model using a regularization parameter.

Fits a Ridge regression model using a regularization parameter.
}
\examples{
X <- matrix(c(1, 2,
              2, 3,
              3, 4,
              4, 5,
              5, 6),
            ncol = 2, byrow = TRUE)
y <- c(2, 3, 5, 7, 8)

ridge_fit(X, y, lambda = 1)

X <- matrix(c(1, 2,
              2, 3,
              3, 4,
              4, 5,
              5, 6),
            ncol = 2, byrow = TRUE)
y <- c(2, 3, 5, 7, 8)

ridge_fit(X, y, lambda = 1)

}
