% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_Spectrum.R
\name{write_Spectrum}
\alias{write_Spectrum}
\title{Write spectrum files to disk}
\usage{
write_Spectrum(x, file, type = c("SPC", "TXT")[1], sep = "\\t",
  gvalue = FALSE, ...)
}
\arguments{
\item{x}{\code{ESR.Spectrum-class} (\strong{required}):
An object of class \code{ESR.Spectrum} that is written to the disk.}

\item{file}{\link{character} (\strong{required}):
A character string naming a file for writing.}

\item{type}{\link{character} (\emph{with default}):
Indicates the file type the data is exported as:
\itemize{
\item \code{"SPC"}: (default) Binary data format as produced by the ESP300-E spectrometer.
Parameters are written to a seperate \code{.PAR} file. See details about encoding.
\item \code{"TXT"}: Plain text file of the spectrum's x,y-values
}}

\item{sep}{\link{character} (\emph{with default}):
The field separator string. Values within each row of x are separated by
this string. Only used if \code{type = 'TXT'}.}

\item{gvalue}{\link{logical} (\emph{with default}):
Export the spectrum with g-values instead of magnetic field values. Requires
all necessary information in the \code{ESR.Spectrum} object (microwave frequency,
magnetic field values, center field, sweep width). Only used if \code{type = 'TXT'}.}

\item{...}{currently not used.}
}
\value{
Write files to the disk, no further output provided.
}
\description{
Export ESR spectrum objects as .SPC/.PAR or .TXT files.
}
\details{
\strong{SPC file format}

If \code{type = 'SPC'} this function will write a binary file as is produced
by a Bruker ESP300-E spectrometer. These files use a big-endian byte order
with 4 bytes per element in the byte stream. This format only stores integer
values, so exporting a floating-point number spectrum might introduce
more or less severe rounding errors. To counteract this, the function
multiplies all intensity values by 1000 before exporting. This preserves
relative differences in spectra, but absolute intensity values may no longer
be interpreted at face value.

The measurement parameters are written to a separate \code{<file>.PAR} file with
two whitespaces as field seperator and macintosh style carriage return
line breaks.
}
\examples{

\dontrun{

## Load example data set
file <- system.file("inst/extdata/quartz.DTA", package = "ESR")

## Import the file
x <- read_Spectrum(file, verbose = FALSE) 

## Export as ESP300-E .SPC/.PAR file set
write_Spectrum(x = x, 
               file = "~/exported.SPC", 
               type = "SPC")

## Export as plain text
write_Spectrum(x = x, 
               file = "~/exported.txt", 
               type = "TXT", 
               gvalue = TRUE, 
               sep = ",")
}

}
\seealso{
\link{writeBin}, \link{write.table}
}
\author{
Christoph Burow, University of Cologne (Germany)
}
