\name{ESR-package}
\alias{ESR-package}
\alias{ESR}
\docType{package}
\title{Collection of functions for Electron Spin Resonance Dating data analysis}
\description{This package provides various functions developed for the purpose of Electron Spin Resonance Dating data analysis.}
\details{
\tabular{ll}{Package: \tab ESR\cr
Type: \tab Package\cr
Version: \tab 1.2.1\cr
Date: \tab 2014-08-14\cr
Author: \tab Christoph Burow \cr
Maintainer: \tab Christoph Burow <christoph.burow@uni-koeln.de>\cr
License: \tab GPL-3\cr}
}
\author{Christoph Burow (University of Cologne, Germany)}

\keyword{ package }


