\name{ExampleData.De}
\alias{ExampleData.De}
\docType{data}
\title{
Synthetic ESR equivalent dose data set
}
\description{
Synthetic ESR equivalent dose data set
}
\usage{ExampleData.De}
\format{
  A data frame with 2 observations on the following 2 variables.
  \describe{
    \item{\code{V1}}{Gamma dose}
    \item{\code{V2}}{ESR intensity}
  }
}
\details{
Just a synthetic data set.
}
\source{
##
}
\references{
##
}
\examples{
data(ExampleData.De)
}
\keyword{datasets}
