% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESR-package.R
\docType{data}
\name{ExampleData.De}
\alias{ExampleData.De}
\title{Synthetic ESR equivalent dose data set}
\format{A data frame with 2 observations on the following 2 variables.
\describe{ \item{V1}{Gamma dose} \item{V2}{ESR intensity} }}
\source{
##
}
\description{
Synthetic ESR equivalent dose data set
}
\details{
Just a synthetic data set.
}
\examples{

data(ExampleData.De, envir = environment())

}
\references{
##
}
\keyword{datasets}

