% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ESR-package.R
\docType{data}
\name{ExampleData.ESRspectra}
\alias{ExampleData.ESRspectra}
\title{ESR spectra of a mollusc and the dpph standard}
\format{A list of ESR spectra where each element contains 1024 observations
on the following 2 variables.}
\source{
In progress
}
\description{
ESR spectra of a mollusc from the foothills of the Alps and the dpph
standard measured at 100 K
}
\details{
In progress
}
\examples{
data(ExampleData.ESRspectra, envir = environment())
}
\references{
In progress
}
\keyword{datasets}

