% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_Spectrum.R
\name{plot_Spectrum}
\alias{plot_Spectrum}
\title{Plot ESR spectra and peak finding}
\usage{
plot_Spectrum(input.data, difference = FALSE, integrate = FALSE,
  smooth.spline = FALSE, smooth.spline.df, smooth.spline.diff.df,
  overlay = TRUE, auto.shift = FALSE, find.peaks = FALSE, peak.range,
  peak.threshold = 10, peak.information = FALSE, manual.peaks = FALSE,
  manual.peaks.roi = FALSE, info = NULL, output.console = TRUE,
  output.plot = TRUE, add = FALSE, ...)
}
\arguments{
\item{input.data}{\code{\link{data.frame}} (\bold{required}): data frame
with two columns for x=magnetic.field or g.value, y=ESR.intensity.}

\item{difference}{\code{\link{logical}} (with default): plot first
derivative of the spectrum}

\item{integrate}{\code{\link{logical}} (with default): plot integrand of the
spectrum}

\item{smooth.spline}{\code{\link{logical}} (with default): fit a cubic
smoothing spline to supplied spectrum.}

\item{smooth.spline.df}{\code{\link{integer}}: desired number of degrees of
freedom}

\item{smooth.spline.diff.df}{\code{\link{integer}}: desired number of
degrees of freedom for splines of the first derivative}

\item{overlay}{\code{\link{logical}} (with default): overlay actual data and
smoothing spline curve in one plot.}

\item{auto.shift}{\code{\link{logical}} (with default): automatically shift
multiple spectra by their maximum peak. This uses smoothing splines for
better results.}

\item{find.peaks}{\code{\link{logical}} (with default): find and plot peaks
(\code{TRUE/FALSE}).}

\item{peak.range}{\code{\link{integer}} (with default): range of magnetic
field intensities or g-values in which peaks are picked from \code{c(from,
to)}. If no values are provide the whole spectrum is analysed.}

\item{peak.threshold}{\code{\link{integer}} (with default): threshold value
specifying the resolution of the peak finding routine (see details).}

\item{peak.information}{\code{\link{logical}} (with default): plot peak
intensity values for peaks found by the automated routine
(\code{TRUE/FALSE}). Applies only when \code{find.peaks = TRUE}.}

\item{manual.peaks}{\code{\link{logical}} (with default): manually identify
xy-coordinates to be used for amplitude calculation (\code{TRUE/FALSE}).}

\item{manual.peaks.roi}{\code{\link{logical}} (with default): define area
for the plot to be zoomed in to allow for a more precise manual peak picking
(\code{TRUE/FALSE}). Applies only when \code{manual.peaks = TRUE}.}

\item{info}{\code{\link{character}}: add information on experimental details
as subtitle}

\item{output.console}{\code{\link{logical}} (with default): print output
(\code{TRUE/FALSE}).}

\item{output.plot}{\code{\link{logical}} (with default): show plot
(\code{TRUE/FALSE}).}

\item{add}{\code{\link{logical}} (with default): whether derivatives and/or
integrands are added to the spectrum or are shown separately
(\code{TRUE/FALSE}).}

\item{\dots}{Further plot arguments to pass.}
}
\value{
Returns terminal output and a plot. In addition, a list is returned
containing the following elements:

\item{data}{list containing the (modified) input data} \item{splines}{list
containing the spline objects} \item{auto.peaks}{data frame containing the
peak information (magnetic field and ESR intensity) found by the peak find
routine.} \item{manual.peaks}{vector containing the peak information
(magnetic field, ESR intensity and amplitude) from the manually identified
peaks.}
}
\description{
Function to plot an ESR spectrum and finding peaks using an automated
routine. Peaks can also be picked manually to calculate the amplitude in
intensity.
}
\details{
\bold{Status} \cr\cr In progress
}
\note{
In progress
}
\examples{
##load example data
data(ExampleData.ESRspectra, envir = environment())

##plot dpph and use the automatic peak finding routine
plot_Spectrum(ExampleData.ESRspectra$dpph, find.peaks = TRUE,
                 peak.range = c(3340,3355),
                 peak.threshold = 10, peak.information = TRUE,
                 output.console = TRUE)

##plot the mollusc (sample Ba01) natural ESR spectrum with a smoothing spline
plot_Spectrum(ExampleData.ESRspectra$Ba01_00,
                 smooth.spline = TRUE,
                 smooth.spline.df = 40,
                 overlay = TRUE)

##plot all ESR spectra of sample Ba01
plot_Spectrum(ExampleData.ESRspectra$Ba01)

##plot all ESR spectra of sample Ba01 and align curves by the max peak
plot_Spectrum(ExampleData.ESRspectra$Ba01,
                 auto.shift = TRUE)

##plot all ESR spectra of sample Ba01, use smoothing splines and
##align curves by the max peak
plot_Spectrum(ExampleData.ESRspectra$Ba01,
                 smooth.spline = TRUE,
                 smooth.spline.df = 40,
                 auto.shift = TRUE,
                 overlay = FALSE)
}
\author{
Christoph Burow, University of Cologne (Germany)
}
\references{
In progress
}
\seealso{
\code{\link{plot}}
}

