% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_Spectrum.R
\name{read_Spectrum}
\alias{read_Spectrum}
\title{Read in single or multiple ESR spectra}
\usage{
read_Spectrum(file, ...)
}
\arguments{
\item{file}{\code{\link{character}} (required): file path or directory where
the spectra files are stored.}

\item{...}{further arguments (e.g., \code{n} to specify the number of datapoints).}
}
\value{
Returns a terminal output. In addition an
\code{\link{R6Class}} object is returned. \cr
}
\description{
Use this function to import a series of associated ESR spectra into R.
}
\details{
This is a wrapper function for \code{read.table}. The function should be
used to read in a series of associated ESR spectrum files. A list with all
spectrum data is returned, which can be passed to \code{plot_Spectrum}
for plotting the spectra.
}
\examples{
# Import ASCII text file
file1 <- system.file("extdata", "coral.txt", package = "ESR")
spec1 <- read_Spectrum(file1)

# Import .zip archive
file2 <- system.file("extdata", "mollusc.zip", package = "ESR")
spec2 <- read_Spectrum(file2)

# Import Bruker ESP300-E raw binary spectrum
file3 <- system.file("extdata", "mollusc.SPC", package = "ESR")
spec3 <- read_Spectrum(file3)

# Import Bruker ELEXSYS500 spectrum (ASCII)
file4 <- system.file("extdata", "dpph.ASC", package = "ESR")
spec4 <- read_Spectrum(file4)

# Import Bruker ELEXSYS500 raw binary spectrum
file5 <- system.file("extdata", "quartz.DTA", package = "ESR")
spec5 <- read_Spectrum(file5)

# Import all example data sets at once by providing only the directory
dir <- system.file("extdata", package = "ESR")
specs <- read_Spectrum(dir)
}
\author{
Christoph Burow, University of Cologne (Germany)
}
\references{
In progress
}
\seealso{
\code{\link{read.table}}, \code{\link{readBin}}, \code{\link{read.csv}}
}

